+++ packages/node/src/utils.ts
@@ -1,7 +1,52 @@
 import { debug, streamToBuffer } from '@vercel/build-utils';
-import { pathToRegexp } from 'path-to-regexp';
+import { pathToRegexp as pathToRegexpCurrent, Key } from 'path-to-regexp';
+import { pathToRegexp as pathToRegexpUpdated } from 'path-to-regexp-updated';
 import type { IncomingMessage } from 'http';
 import { extname } from 'path';
+
+// run the updated version of path-to-regexp, compare the results, and log if different
+function pathToRegexp(
+  callerId: string,
+  path: string,
+  keys?: Key[],
+  options?: { strict: boolean; sensitive: boolean; delimiter: string }
+) {
+  const currentRegExp = pathToRegexpCurrent(path, keys, options);
+
+  try {
+    const currentKeys = keys;
+    const newKeys: Key[] = [];
+    const newRegExp = pathToRegexpUpdated(path, newKeys, options);
+
+    // FORCE_PATH_TO_REGEXP_LOG can be used to force these logs to render
+    // for verification that they show up in the build logs as expected
+
+    const isDiffRegExp = currentRegExp.toString() !== newRegExp.toString();
+    if (process.env.FORCE_PATH_TO_REGEXP_LOG || isDiffRegExp) {
+      const message = JSON.stringify({
+        path,
+        currentRegExp: currentRegExp.toString(),
+        newRegExp: newRegExp.toString(),
+      });
+      console.error(`[vc] PATH TO REGEXP PATH DIFF @ #${callerId}: ${message}`);
+    }
+
+    const isDiffKeys = keys?.toString() !== newKeys?.toString();
+    if (process.env.FORCE_PATH_TO_REGEXP_LOG || isDiffKeys) {
+      const message = JSON.stringify({
+        isDiffKeys,
+        currentKeys,
+        newKeys,
+      });
+      console.error(`[vc] PATH TO REGEXP KEYS DIFF @ #${callerId}: ${message}`);
+    }
+  } catch (err) {
+    const error = err as Error;
+    console.error(`[vc] PATH TO REGEXP ERROR @ #${callerId}: ${error.message}`);
+  }
+
+  return currentRegExp;
+}
 
 // When exiting this process, wait for Vercel Function server to finish
 // all its work, especially waitUntil promises before exiting this process.
@@ -47,9 +92,9 @@ function getRegExpFromMatcher(
     );
   }
 
-  const regExps = [pathToRegexp(matcher).source];
+  const regExps = [pathToRegexp('316', matcher).source];
   if (matcher === '/' && !allMatchers.includes('/index')) {
-    regExps.push(pathToRegexp('/index').source);
+    regExps.push(pathToRegexp('491', '/index').source);
   }
   return regExps;
 }