+++ packages/python/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @vercel/python
 
+## 4.7.1
+
+### Patch Changes
+
+- Remove support for VERCEL_IPC_FD ([#12908](https://github.com/vercel/vercel/pull/12908))
+
+- Add `supportsResponseStreaming` to build output ([#12884](https://github.com/vercel/vercel/pull/12884))
+
 ## 4.7.0
 
 ### Minor Changes